Fix tests by removing bad mocks and general refactoring
