+++ src/main/java/org/damap/base/rest/ProjectService.java
@@ -5,6 +5,7 @@ import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.enterprise.inject.Typed;
 import jakarta.inject.Inject;
 import java.util.List;
+import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
@@ -59,7 +60,7 @@ public final class ProjectService implements ProjectServiceProvider {
                       // implement this function correctly.
                       return true;
                     }
-                    return configId.equals(selectedService);
+                    return Objects.equals(configId, selectedService);
                   })
               .findFirst()
               .orElse(null);
