+++ src/test/java/org/damap/base/TestSetup.java
@@ -4,10 +4,13 @@ import static org.mockito.ArgumentMatchers.any;
 
 import io.quarkus.test.InjectMock;
 import io.quarkus.test.junit.QuarkusTest;
+import io.quarkus.test.junit.TestProfile;
 import jakarta.inject.Inject;
+import org.damap.base.integration.mock.MockProjectServiceImpl;
 import org.damap.base.integration.mock.MockUniversityPersonServiceImpl;
 import org.damap.base.integration.orcid.ORCIDMapper;
 import org.damap.base.integration.orcid.ORCIDPersonServiceImpl;
+import org.damap.base.integration.pure.PureIntegrationTestProfile;
 import org.damap.base.rest.dmp.domain.DmpDO;
 import org.damap.base.security.SecurityService;
 import org.damap.base.util.TestDOFactory;
@@ -16,7 +19,6 @@ import org.mockito.Mockito;
 
 // Common config for test case setup
 /** TestSetup class. */
-@QuarkusTest
 public class TestSetup {
   @Inject TestDOFactory testDOFactory;
 
@@ -26,9 +28,11 @@ public class TestSetup {
 
   @InjectMock protected ORCIDPersonServiceImpl orcidPersonServiceImpl;
 
+  @InjectMock MockProjectServiceImpl projectService;
+
   protected DmpDO dmpDO;
 
-  /** setup. */
+  /** @BeforeEach in an extending class overrides this, use super.setup() to circumvent */
   @BeforeEach
   public void setup() {
     Mockito.when(securityService.getUserId()).thenReturn("012345");
@@ -36,9 +40,9 @@ public class TestSetup {
     Mockito.when(personService.read(any(String.class)))
         .thenReturn(testDOFactory.getTestContributorDO());
     Mockito.when(orcidPersonServiceImpl.read(any(String.class)))
-        .thenReturn(testDOFactory.getTestContributorDO());
-    Mockito.when(orcidPersonServiceImpl.read(any(String.class)))
         .thenReturn(ORCIDMapper.mapRecordEntityToPersonDO(testDOFactory.getORCIDTestRecord()));
+    Mockito.when(projectService.getProjectLeader(any()))
+            .thenReturn(testDOFactory.getTestContributorDO());
     dmpDO = testDOFactory.getOrCreateTestDmpDO();
   }
 }
