+++ src/test/java/org/damap/base/conversion/ExportFWFTemplateTest.java
@@ -2,9 +2,11 @@ package org.damap.base.conversion;
 
 import io.quarkus.test.InjectMock;
 import io.quarkus.test.junit.QuarkusTest;
+import io.quarkus.test.junit.TestProfile;
 import io.quarkus.test.security.TestSecurity;
 import jakarta.inject.Inject;
 import org.apache.poi.xwpf.usermodel.XWPFDocument;
+import org.damap.base.TestProfiles;
 import org.damap.base.TestSetup;
 import org.damap.base.integration.mock.MockProjectServiceImpl;
 import org.damap.base.rest.dmp.domain.DmpDO;
@@ -13,13 +15,12 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 @QuarkusTest
+@TestProfile(TestProfiles.DefaultProfile.class)
 class ExportFWFTemplateTest extends TestSetup {
 
   @Inject ExportFWFTemplate exportFWFTemplate;
 
   @Inject TestDOFactory testDOFactory;
-
-  @InjectMock MockProjectServiceImpl mockProjectService;
 
   @Test
   @TestSecurity(authorizationEnabled = false)
