+++ src/test/java/org/damap/base/domain/DatasetSizeRangeTest.java
@@ -1,10 +1,13 @@
 package org.damap.base.domain;
 
 import io.quarkus.test.junit.QuarkusTest;
+import io.quarkus.test.junit.TestProfile;
+import org.damap.base.TestProfiles;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 @QuarkusTest
+@TestProfile(TestProfiles.DefaultProfile.class)
 class DatasetSizeRangeTest {
 
   @Test
