+++ src/test/java/org/damap/base/rest/DmpDocumentResourceTest.java
@@ -6,8 +6,11 @@ import static org.mockito.ArgumentMatchers.anyString;
 import io.quarkus.test.InjectMock;
 import io.quarkus.test.common.http.TestHTTPEndpoint;
 import io.quarkus.test.junit.QuarkusTest;
+import io.quarkus.test.junit.TestProfile;
 import io.quarkus.test.security.TestSecurity;
 import jakarta.inject.Inject;
+import org.damap.base.TestProfiles;
+import org.damap.base.TestSetup;
 import org.damap.base.integration.mock.MockProjectServiceImpl;
 import org.damap.base.rest.dmp.domain.DmpDO;
 import org.damap.base.security.SecurityService;
@@ -18,21 +21,8 @@ import org.mockito.Mockito;
 
 @QuarkusTest
 @TestHTTPEndpoint(DmpDocumentResource.class)
-class DmpDocumentResourceTest {
-
-  @Inject TestDOFactory testDOFactory;
-
-  @InjectMock SecurityService securityService;
-
-  @InjectMock MockProjectServiceImpl mockProjectService;
-
-  /** setup. */
-  @BeforeEach
-  public void setup() {
-    Mockito.when(securityService.getUserId()).thenReturn("012345");
-    Mockito.when(securityService.getUserName()).thenReturn("testUser");
-    Mockito.when(mockProjectService.read(anyString())).thenReturn(testDOFactory.getTestProjectDO());
-  }
+@TestProfile(TestProfiles.DefaultProfile.class)
+class DmpDocumentResourceTest extends TestSetup {
 
   @Test
   void testExportTemplateEndpoint_Invalid() {
@@ -48,35 +38,30 @@ class DmpDocumentResourceTest {
   @Test
   @TestSecurity(user = "userJwt", roles = "user")
   void testExportTemplateEndpoint_Valid() {
-    DmpDO dmpDO = testDOFactory.getOrCreateTestDmpDO();
     given().when().get("/" + dmpDO.getId()).then().statusCode(200);
   }
 
   @Test
   @TestSecurity(user = "userJwt", roles = "user")
   void testExportTemplateEndpointWithTemplateTypeFWF_Valid() {
-    DmpDO dmpDO = testDOFactory.getOrCreateTestDmpDO();
     given().when().get("/" + dmpDO.getId() + "?template=FWF").then().statusCode(200);
   }
 
   @Test
   @TestSecurity(user = "userJwt", roles = "user")
   void testExportTemplateEndpointWithTemplateTypeScienceEurope_Valid() {
-    DmpDO dmpDO = testDOFactory.getOrCreateTestDmpDO();
     given().when().get("/" + dmpDO.getId() + "?template=SCIENCE_EUROPE").then().statusCode(200);
   }
 
   @Test
   @TestSecurity(user = "userJwt", roles = "user")
   void testExportTemplateEndpointWithTemplateType_Invalid() {
-    DmpDO dmpDO = testDOFactory.getOrCreateTestDmpDO();
     given().when().get("/" + dmpDO.getId() + "?template=invalid").then().statusCode(404);
   }
 
   @Test
   @TestSecurity(user = "userJwt", roles = "user")
   void testExportEndpointForPDFWithTemplateTypeScienceEurope_Valid() {
-    DmpDO dmpDO = testDOFactory.getOrCreateTestDmpDO();
     given()
         .when()
         .get("/" + dmpDO.getId() + "/export?template=SCIENCE_EUROPE&fileType=docx")
@@ -87,7 +72,6 @@ class DmpDocumentResourceTest {
   @Test
   @TestSecurity(user = "userJwt", roles = "user")
   void testExportEndpointForPDFWithTemplateTypeHorizonEurope_Valid() {
-    DmpDO dmpDO = testDOFactory.getOrCreateTestDmpDO();
     given()
         .when()
         .get("/" + dmpDO.getId() + "/export?template=HORIZON_EUROPE&fileType=docx")
@@ -98,7 +82,6 @@ class DmpDocumentResourceTest {
   @Test
   @TestSecurity(user = "userJwt", roles = "user")
   void testExportEndpointForPDFWithTemplateTypeFWF_Valid() {
-    DmpDO dmpDO = testDOFactory.getOrCreateTestDmpDO();
     given()
         .when()
         .get("/" + dmpDO.getId() + "/export?template=FWF&fileType=docx")
