+++ src/test/java/org/damap/base/rest/InvenioDamapResourceDisabledTest.java
@@ -4,11 +4,14 @@ import static io.restassured.RestAssured.given;
 
 import io.quarkus.test.common.http.TestHTTPEndpoint;
 import io.quarkus.test.junit.QuarkusTest;
+import io.quarkus.test.junit.TestProfile;
+import org.damap.base.TestProfiles;
 import org.damap.base.rest.invenio_damap.InvenioDAMAPResource;
 import org.junit.jupiter.api.Test;
 
 @QuarkusTest
 @TestHTTPEndpoint(InvenioDAMAPResource.class)
+@TestProfile(TestProfiles.DefaultProfile.class)
 class InvenioDamapResourceDisabledTest {
   @Test
   void givenConfigNotEnabled_thenNoEndpointsFound() {
