+++ src/test/java/org/damap/base/rest/dmp/DmpConsistencyUtilityTest.java
@@ -6,7 +6,10 @@ import io.quarkus.test.junit.QuarkusTest;
 import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
+
+import io.quarkus.test.junit.TestProfile;
 import lombok.extern.jbosslog.JBossLog;
+import org.damap.base.TestProfiles;
 import org.damap.base.enums.*;
 import org.damap.base.rest.dmp.domain.*;
 import org.damap.base.rest.dmp.service.DmpConsistencyUtility;
@@ -14,6 +17,7 @@ import org.junit.jupiter.api.Test;
 
 @QuarkusTest
 @JBossLog
+@TestProfile(TestProfiles.DefaultProfile.class)
 class DmpConsistencyUtilityTest {
 
   @Test
