+++ src/test/java/org/damap/base/rest/madmp/MaDmpServiceTest.java
@@ -1,8 +1,11 @@
 package org.damap.base.rest.madmp;
 
 import io.quarkus.test.junit.QuarkusTest;
+import io.quarkus.test.junit.TestProfile;
 import io.quarkus.test.security.TestSecurity;
 import jakarta.inject.Inject;
+import org.damap.base.TestProfiles;
+import org.damap.base.TestSetup;
 import org.damap.base.rest.dmp.mapper.MapperService;
 import org.damap.base.rest.madmp.dto.Dmp;
 import org.damap.base.rest.madmp.mapper.MaDmpMapper;
@@ -11,7 +14,8 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 @QuarkusTest
-class MaDmpServiceTest {
+@TestProfile(TestProfiles.DefaultProfile.class)
+class MaDmpServiceTest extends TestSetup {
 
   @Inject TestDOFactory testDOFactory;
